requirements.txt
Flask>=3.0.0
edge-tts>=0.14.0
striprtf==0.0.26
python-dotenv==1.0.1
Flask-WTF==1.2.1
Werkzeug>=3.0.0
Flask-Limiter==3.5.0
gunicorn>=21.2.0
